feat(tvf): support for testing usage of tvf in smoke test
The check is a bit different from table and views. It has to verify that the schema
matches and the params. The schema tests is done through a CTE and the parameters
through a UDF that always return true. The query CTE body needs to include a reference
to both of them.