feat: lift BQSqlFrag into a higher structure
Validate and lift the fragments into either a pure or session query. A pure
query is a self-contained query and the session query requires a session to
be executed. A pure query can be used in views but a session query might not.
A session query can contain parameters and temporary UDF etc.